+++ cartridge/twophase.lua
@@ -506,10 +506,20 @@ local function _force_reapply(uuids)
     -- Prepare a server group to be configured
     local uri_list = {}
     local refined_uri_list = topology.refine_servers_uri(current_topology)
-    for _, uuid, _ in fun.filter(topology.not_disabled, current_topology.servers) do
-        if utils.table_find(uuids, uuid) then
-            table.insert(uri_list, refined_uri_list[uuid])
+    for _, uuid in ipairs(uuids) do
+        local srv = current_topology.servers[uuid]
+        if not srv then
+            return nil, ForceReapplyError:new(
+                'Server %s not in clusterwide config', uuid
+            )
+        elseif not topology.not_disabled(uuid, srv) then
+            return nil, ForceReapplyError:new(
+                'Server %s is disabled, not suitable' ..
+                ' for reapplying config', uuid
+            )
         end
+
+        table.insert(uri_list, refined_uri_list[uuid])
     end
 
     do